+++ src/components/common/navbar/LoginBox.jsx
@@ -2,7 +2,7 @@ import { Link } from "react-router-dom";
 
 const LoginBox = () => {
   return (
-    <div className="flex flex-col md:flex-row gap-6 px-3 mt-0 ">
+    <div className="flex flex-col md:flex-row gap-6 px-3 mt-0 items-center ">
       <Link to="/signin" className="text-mCream  hover:text-mYellow text-lg">
         로그인
       </Link>
